+++ tango/core/rt/compiler/ldc/rt/lifetime.d
@@ -1127,7 +1127,7 @@ extern (C) byte[] _d_arraycatnT(TypeInfo ti, uint n, ...)
     }
 
     byte[] result;
-    *cast(int *)&result = length;       // jam length
+    *cast(size_t *)&result = length;       // jam length
     (cast(void **)&result)[1] = a;      // jam ptr
     return result;
 }
